requirements.txt
discord.py
openai
python-dotenv
yt-dlp
PyNaCl